refactor: badge 컴포넌트 index 파일 생성 및 export 선언
